Expose typed inbound LXMF metadata
