+++ crates/libs/lxmf-sdk/src/app/events.rs
@@ -1,6 +1,11 @@
 #[cfg(feature = "sdk-async")]
 use crate::event::{EventBatch as RawEventBatch, EventSubscription, SdkEvent};
 use crate::event::{Severity as RawSeverity, SubscriptionStart as RawSubscriptionStart};
+#[path = "events_details.rs"]
+mod events_details;
+pub use events_details::{
+    DeliveryLifecycleDetails, InboundDropDetails, InboundMessageDetails, StreamGapDetails,
+};
 use serde::{Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 use std::collections::BTreeMap;
@@ -57,61 +62,6 @@ impl From<RawSubscriptionStart> for SubscriptionStart {
             RawSubscriptionStart::Snapshot => Self::Snapshot,
         }
     }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-#[non_exhaustive]
-pub struct StreamGapDetails {
-    pub expected_seq_no: Option<u64>,
-    pub observed_seq_no: Option<u64>,
-    pub dropped_count: u64,
-    pub recovery_required: bool,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-#[non_exhaustive]
-pub struct InboundMessageDetails {
-    pub message_id: Option<String>,
-    pub source_hash: Option<String>,
-    pub destination_hash: Option<String>,
-    pub delivery_kind: Option<String>,
-    pub lxmf_bytes_hex: Option<String>,
-    pub receipt_status: Option<String>,
-    pub signature_checked: Option<bool>,
-    pub signature_status: Option<String>,
-    pub stamp_status: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-#[non_exhaustive]
-pub struct InboundDropDetails {
-    pub reason: Option<String>,
-    pub delivery_kind: Option<String>,
-    pub raw_destination_hash: Option<String>,
-    pub resolved_destination_hash: Option<String>,
-    pub source_hash: Option<String>,
-    pub destination_hash: Option<String>,
-    pub dropped_message_id: Option<String>,
-    pub payload_mode: Option<String>,
-    pub bytes_len: Option<u64>,
-    pub detail: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-#[non_exhaustive]
-pub struct DeliveryLifecycleDetails {
-    pub state: Option<String>,
-    pub from: Option<String>,
-    pub to: Option<String>,
-    pub receipt_status: Option<String>,
-    pub delivery_kind: Option<String>,
-    pub packet_hash: Option<String>,
-    pub resource_hash: Option<String>,
-    pub peer: Option<String>,
-    pub method: Option<String>,
-    pub bytes: Option<u64>,
-    pub link_id: Option<String>,
-    pub reason: Option<String>,
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
@@ -186,11 +136,11 @@ pub struct EventBatch {
 impl Event {
     pub fn inbound_message_details(&self) -> Option<InboundMessageDetails> {
         matches!(self.kind, EventKind::InboundMessageReceived)
-            .then(|| inbound_message_details(&self.details))
+            .then(|| events_details::inbound_message_details(&self.details))
     }
     pub fn inbound_drop_details(&self) -> Option<InboundDropDetails> {
         matches!(self.kind, EventKind::InboundMessageDropped)
-            .then(|| inbound_drop_details(&self.details))
+            .then(|| events_details::inbound_drop_details(&self.details))
     }
     pub fn delivery_lifecycle_details(&self) -> Option<DeliveryLifecycleDetails> {
         matches!(
@@ -202,79 +152,7 @@ impl Event {
                 | EventKind::MessageFailed
                 | EventKind::MessageCancelled
         )
-        .then(|| delivery_lifecycle_details(&self.details))
-    }
-}
-fn json_str(value: &JsonValue, key: &str) -> Option<String> {
-    value.get(key)?.as_str().map(ToOwned::to_owned)
-}
-
-fn nested_json_str(value: &JsonValue, path: &[&str]) -> Option<String> {
-    let mut current = value;
-    for key in path {
-        current = current.get(*key)?;
-    }
-    current.as_str().map(ToOwned::to_owned)
-}
-
-fn nested_json_bool(value: &JsonValue, path: &[&str]) -> Option<bool> {
-    let mut current = value;
-    for key in path {
-        current = current.get(*key)?;
-    }
-    current.as_bool()
-}
-
-fn inbound_message_details(payload: &JsonValue) -> InboundMessageDetails {
-    let message = payload.get("message").unwrap_or(payload);
-    InboundMessageDetails {
-        message_id: json_str(message, "id").or_else(|| json_str(payload, "message_id")),
-        source_hash: json_str(message, "source").or_else(|| json_str(payload, "source_hash")),
-        destination_hash: json_str(message, "destination")
-            .or_else(|| json_str(payload, "destination_hash")),
-        delivery_kind: json_str(payload, "delivery_kind"),
-        lxmf_bytes_hex: json_str(payload, "lxmf_bytes_hex"),
-        receipt_status: json_str(message, "receipt_status")
-            .or_else(|| json_str(payload, "receipt_status")),
-        signature_checked: nested_json_bool(message, &["fields", "_lxmf", "signature_checked"]),
-        signature_status: nested_json_str(message, &["fields", "_lxmf", "signature_status"]),
-        stamp_status: nested_json_str(message, &["fields", "_lxmf", "stamp_status"]),
-    }
-}
-
-fn inbound_drop_details(payload: &JsonValue) -> InboundDropDetails {
-    InboundDropDetails {
-        reason: json_str(payload, "reason"),
-        delivery_kind: json_str(payload, "delivery_kind"),
-        raw_destination_hash: json_str(payload, "raw_destination_hash"),
-        resolved_destination_hash: json_str(payload, "resolved_destination_hash"),
-        source_hash: json_str(payload, "source_hash"),
-        destination_hash: json_str(payload, "destination_hash"),
-        dropped_message_id: json_str(payload, "dropped_message_id"),
-        payload_mode: json_str(payload, "payload_mode"),
-        bytes_len: payload.get("bytes_len").and_then(JsonValue::as_u64),
-        detail: json_str(payload, "detail"),
-    }
-}
-
-fn delivery_lifecycle_details(payload: &JsonValue) -> DeliveryLifecycleDetails {
-    let message = payload.get("message").unwrap_or(payload);
-    DeliveryLifecycleDetails {
-        state: json_str(payload, "state")
-            .or_else(|| normalized_receipt_state(payload).ok().flatten()),
-        from: json_str(payload, "from"),
-        to: json_str(payload, "to"),
-        receipt_status: json_str(message, "receipt_status")
-            .or_else(|| json_str(payload, "receipt_status"))
-            .or_else(|| json_str(payload, "status")),
-        delivery_kind: json_str(payload, "delivery_kind"),
-        packet_hash: json_str(payload, "packet_hash"),
-        resource_hash: json_str(payload, "resource_hash"),
-        peer: json_str(payload, "peer").or_else(|| json_str(payload, "peer_id")),
-        method: json_str(payload, "method"),
-        bytes: payload.get("bytes").and_then(JsonValue::as_u64),
-        link_id: json_str(payload, "link_id"),
-        reason: json_str(payload, "reason").or_else(|| json_str(payload, "detail")),
+        .then(|| events_details::delivery_lifecycle_details(&self.details))
     }
 }
 
@@ -287,19 +165,6 @@ fn payload_state(payload: &JsonValue, key: &str) -> Result<Option<String>, &'sta
             .ok_or("payload field is not a string")
             .map(|s| Some(s.trim().to_ascii_lowercase())),
     }
-}
-
-fn normalized_receipt_state(payload: &JsonValue) -> Result<Option<String>, &'static str> {
-    let status_val = payload
-        .get("message")
-        .and_then(|message| message.get("receipt_status").or_else(|| message.get("status")))
-        .or_else(|| payload.get("receipt_status"))
-        .or_else(|| payload.get("status"))
-        .or_else(|| payload.get("state"))
-        .or_else(|| payload.get("receipt").and_then(|receipt| receipt.get("status")));
-    let Some(status_val) = status_val else { return Ok(None) };
-    let status = status_val.as_str().ok_or("receipt status is not a string")?;
-    Ok(Some(status.split(':').next().unwrap_or(status).trim().to_ascii_lowercase()))
 }
 
 #[cfg(feature = "sdk-async")]
@@ -444,7 +309,7 @@ pub fn map_sdk_event(event: SdkEvent, profile_id: &str) -> Event {
         "sdk_security_rate_limited" => EventKind::SecurityActionRequired,
         "runtime_shutdown_requested" => EventKind::RuntimeStopped,
         "outbound" | "receipt" => map_delivery_state(
-            normalized_receipt_state(&event.payload)
+            events_details::normalized_receipt_state(&event.payload)
                 .ok()
                 .flatten()
                 .unwrap_or_else(|| "unknown".to_owned())
